requirements.txt
requests==2.31.0
pandas==2.2.3
pymysql==1.1.2
SQLAlchemy==2.0.49
python-dotenv==1.2.1
Flask==3.0.0
jieba==0.42.1
snownlp==0.12.3
wordcloud==1.9.6
matplotlib>=3.10.0
